+++ mm/2s2h/resource/type/TextMM.h
@@ -13,7 +13,7 @@ class MessageEntryMM {
     uint16_t id;
     uint8_t textboxType;
     uint8_t textboxYPos;
-    uint8_t icon;
+    uint16_t icon;
     uint16_t nextMessageID;
     uint16_t firstItemCost;
     uint16_t secondItemCost;
